requirements.txt
GitPython
watchdog
tk
keyring
schedule